+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.3
 * feature: `props.sortConfig.defaultSort` 控制默认排序。
+* bugfix: 修复 `props.dataSource` 为 `shallowRef` 时，高亮行异常的问题。
 
 ## 0.2.2
 * feature: `props.sortConfig` 配置排序规则排序。